backend/requirements.txt
ansi2html==1.9.1
asttokens==2.4.1
attrs==23.1.0
blinker==1.7.0
category-encoders==2.6.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.6
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
Flask==3.0.0
fonttools==4.46.0
idna==3.6
imbalanced-learn==0.11.0
importlib-metadata==7.0.0
ipython==8.18.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_core==5.5.0
jupyterlab-widgets==3.0.9
kaleido==0.2.1
kiwisolver==1.4.5
lightgbm==4.1.0
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.6.0
matplotlib-inline==0.1.6
nbformat==5.9.2
neo4j==5.15.0
nest-asyncio==1.5.8
numba==0.58.1
numpy==1.25.2
orjson==3.9.10
packaging==23.2
pandas==1.5.3
parso==0.8.3
patsy==0.5.4
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
plotly==5.18.0
plotly-resampler==0.9.1
pmdarima==2.0.4
prompt-toolkit==3.0.43
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycaret==3.2.0
Pygments==2.17.2
pyod==1.1.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
retrying==1.3.4
rpds-py==0.13.2
schemdraw==0.15
scikit-base==0.5.2
scikit-learn==1.2.2
scikit-plot==0.3.7
scipy==1.10.1
six==1.16.0
sktime==0.21.1
stack-data==0.6.3
statsmodels==0.14.1
tbats==1.1.3
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
trace-updater==0.0.9.1
traitlets==5.14.0
tsdownsample==0.1.2
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
wurlitzer==3.0.3
xxhash==3.4.1
yellowbrick==1.5
zipp==3.17.0
